int CBPStructure progressing.
git-svn-id: svn://127.0.0.1@189 2ece6d47-a862-c445-b01b-7049066ce7d0

+++ JMM/ldecod/inc/global.h
@@ -206,9 +206,9 @@ typedef struct DecRefPicMarking_s
 //! cbp structure
 typedef struct cbp_s
 {
-  int64         blk     ;
-  int64         bits    ;
-  int64         bits_8x8;
+  int         blk     ;
+  int         bits    ;
+  int         bits_8x8;
 } CBPStructure;
 
 //! Macroblock
